Refactor frontend into modules and rename entrypoint
