+++ lockstep_compiler/__init__.py
@@ -1,0 +1,18 @@
+from lockstep_compiler.diagnostics import (
+    LockstepCompileError,
+    LockstepCompileResult,
+    LockstepDiagnostic,
+    ParseErrorCollector,
+    normalize_diagnostics,
+)
+from lockstep_compiler.visitors import LockstepDebugVisitor, LockstepSemanticValidator
+
+__all__ = [
+    "LockstepCompileError",
+    "LockstepCompileResult",
+    "LockstepDiagnostic",
+    "ParseErrorCollector",
+    "normalize_diagnostics",
+    "LockstepDebugVisitor",
+    "LockstepSemanticValidator",
+]
