+++ tests/test_lockstep_frontend.py
@@ -14,7 +14,7 @@ if str(REPO_ROOT) not in sys.path:
 
 
 def _install_fake_generated_modules(monkeypatch):
-    """Install minimal ANTLR-generated modules so debug_compiler can import."""
+    """Install minimal ANTLR-generated modules so lockstep_frontend can import."""
 
     lexer_module = types.ModuleType("LockstepLexer")
 
@@ -119,10 +119,13 @@ def _install_fake_generated_modules(monkeypatch):
 
 
 @pytest.fixture
-def debug_compiler_module(monkeypatch):
+def lockstep_frontend_module(monkeypatch):
     _install_fake_generated_modules(monkeypatch)
-    sys.modules.pop("debug_compiler", None)
-    return importlib.import_module("debug_compiler")
+    sys.modules.pop("lockstep_frontend", None)
+    sys.modules.pop("lockstep_compiler", None)
+    sys.modules.pop("lockstep_compiler.visitors", None)
+    sys.modules.pop("lockstep_compiler.diagnostics", None)
+    return importlib.import_module("lockstep_frontend")
 
 
 def _diagnostic(module, line, column, message, *, severity="error", code="LCK001"):
@@ -136,14 +139,14 @@ def _diagnostic(module, line, column, message, *, severity="error", code="LCK001
     )
 
 
-def test_lockstep_compile_error_formats_singular_and_plural(debug_compiler_module):
-    one = debug_compiler_module.LockstepCompileError(
-        [_diagnostic(debug_compiler_module, 1, 1, "oops")]
+def test_lockstep_compile_error_formats_singular_and_plural(lockstep_frontend_module):
+    one = lockstep_frontend_module.LockstepCompileError(
+        [_diagnostic(lockstep_frontend_module, 1, 1, "oops")]
     )
-    many = debug_compiler_module.LockstepCompileError(
+    many = lockstep_frontend_module.LockstepCompileError(
         [
-            _diagnostic(debug_compiler_module, 1, 1, "oops"),
-            _diagnostic(debug_compiler_module, 2, 4, "bad"),
+            _diagnostic(lockstep_frontend_module, 1, 1, "oops"),
+            _diagnostic(lockstep_frontend_module, 2, 4, "bad"),
         ]
     )
 
@@ -154,12 +157,12 @@ def test_lockstep_compile_error_formats_singular_and_plural(debug_compiler_modul
     )
 
 
-def test_parse_error_collector_captures_diagnostic(debug_compiler_module):
-    collector = debug_compiler_module.ParseErrorCollector()
+def test_parse_error_collector_captures_diagnostic(lockstep_frontend_module):
+    collector = lockstep_frontend_module.ParseErrorCollector()
     collector.syntaxError(None, None, 12, 7, "unexpected token", None)
 
     assert collector.errors == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="error",
             code="LCK001",
             message="unexpected token",
@@ -171,7 +174,7 @@ def test_parse_error_collector_captures_diagnostic(debug_compiler_module):
 
 
 def test_compile_lockstep_raises_when_parser_reports_errors(
-    debug_compiler_module, monkeypatch
+    lockstep_frontend_module, monkeypatch
 ):
     class FailingParser:
         def __init__(self, stream):
@@ -189,15 +192,15 @@ def test_compile_lockstep_raises_when_parser_reports_errors(
             return object()
 
     monkeypatch.setattr(
-        debug_compiler_module, "CommonTokenStream", lambda lexer: object()
+        lockstep_frontend_module, "CommonTokenStream", lambda lexer: object()
     )
-    monkeypatch.setattr(debug_compiler_module, "LockstepParser", FailingParser)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepParser", FailingParser)
 
-    with pytest.raises(debug_compiler_module.LockstepCompileError) as exc_info:
-        debug_compiler_module.compile_lockstep("pipeline P { }")
+    with pytest.raises(lockstep_frontend_module.LockstepCompileError) as exc_info:
+        lockstep_frontend_module.compile_lockstep("pipeline P { }")
 
     assert exc_info.value.errors == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="error",
             code="LCK001",
             message="mismatched input",
@@ -209,7 +212,7 @@ def test_compile_lockstep_raises_when_parser_reports_errors(
     assert exc_info.value.diagnostics == exc_info.value.errors
 
 
-def test_compile_lockstep_visits_tree_on_success(debug_compiler_module, monkeypatch):
+def test_compile_lockstep_visits_tree_on_success(lockstep_frontend_module, monkeypatch):
     class SuccessParser:
         def __init__(self, stream):
             self._listeners = []
@@ -237,7 +240,7 @@ def test_compile_lockstep_visits_tree_on_success(debug_compiler_module, monkeypa
             self.uniforms = [{"name": "dt", "type": "float", "initializer": "0.016"}]
             self.bind_routes = ["final=ApplyGravity(raw,final,energy,dt);"]
             self.diagnostics = [
-                debug_compiler_module.LockstepDiagnostic(
+                lockstep_frontend_module.LockstepDiagnostic(
                     severity="warning",
                     code="LCK203",
                     message="Stream 'raw' is redeclared.",
@@ -251,12 +254,12 @@ def test_compile_lockstep_visits_tree_on_success(debug_compiler_module, monkeypa
             visited["tree"] = tree
 
     monkeypatch.setattr(
-        debug_compiler_module, "CommonTokenStream", lambda lexer: object()
+        lockstep_frontend_module, "CommonTokenStream", lambda lexer: object()
     )
-    monkeypatch.setattr(debug_compiler_module, "LockstepParser", SuccessParser)
-    monkeypatch.setattr(debug_compiler_module, "LockstepDebugVisitor", SpyVisitor)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepParser", SuccessParser)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepDebugVisitor", SpyVisitor)
 
-    result = debug_compiler_module.compile_lockstep("pipeline P { }", verbose=False)
+    result = lockstep_frontend_module.compile_lockstep("pipeline P { }", verbose=False)
 
     assert visited["tree"] == "TREE"
     assert result.parse_tree == "TREE"
@@ -271,7 +274,7 @@ def test_compile_lockstep_visits_tree_on_success(debug_compiler_module, monkeypa
         "bind_routes": ["final=ApplyGravity(raw,final,energy,dt);"],
     }
     assert result.diagnostics == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="warning",
             code="LCK203",
             message="Stream 'raw' is redeclared.",
@@ -283,7 +286,7 @@ def test_compile_lockstep_visits_tree_on_success(debug_compiler_module, monkeypa
 
 
 def test_compile_lockstep_runs_semantic_validation_phase(
-    debug_compiler_module, monkeypatch
+    lockstep_frontend_module, monkeypatch
 ):
     class SuccessParser:
         def __init__(self, stream):
@@ -299,14 +302,14 @@ def test_compile_lockstep_runs_semantic_validation_phase(
             return "TREE"
 
     monkeypatch.setattr(
-        debug_compiler_module, "CommonTokenStream", lambda lexer: object()
+        lockstep_frontend_module, "CommonTokenStream", lambda lexer: object()
     )
-    monkeypatch.setattr(debug_compiler_module, "LockstepParser", SuccessParser)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepParser", SuccessParser)
     monkeypatch.setattr(
-        debug_compiler_module,
+        lockstep_frontend_module,
         "validate_semantics",
         lambda parse_tree: [
-            debug_compiler_module.LockstepDiagnostic(
+            lockstep_frontend_module.LockstepDiagnostic(
                 severity="info",
                 code="LCK301",
                 message=f"validated {parse_tree}",
@@ -317,10 +320,10 @@ def test_compile_lockstep_runs_semantic_validation_phase(
         ],
     )
 
-    result = debug_compiler_module.compile_lockstep("pipeline P { }", verbose=False)
+    result = lockstep_frontend_module.compile_lockstep("pipeline P { }", verbose=False)
 
     assert result.diagnostics == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="info",
             code="LCK301",
             message="validated TREE",
@@ -332,7 +335,7 @@ def test_compile_lockstep_runs_semantic_validation_phase(
 
 
 def test_compile_lockstep_raises_for_semantic_errors(
-    debug_compiler_module, monkeypatch
+    lockstep_frontend_module, monkeypatch
 ):
     class SuccessParser:
         def __init__(self, stream):
@@ -348,14 +351,14 @@ def test_compile_lockstep_raises_for_semantic_errors(
             return "TREE"
 
     monkeypatch.setattr(
-        debug_compiler_module, "CommonTokenStream", lambda lexer: object()
+        lockstep_frontend_module, "CommonTokenStream", lambda lexer: object()
     )
-    monkeypatch.setattr(debug_compiler_module, "LockstepParser", SuccessParser)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepParser", SuccessParser)
     monkeypatch.setattr(
-        debug_compiler_module,
+        lockstep_frontend_module,
         "validate_semantics",
         lambda _parse_tree: [
-            debug_compiler_module.LockstepDiagnostic(
+            lockstep_frontend_module.LockstepDiagnostic(
                 severity="error",
                 code="LCK401",
                 message="semantic problem",
@@ -366,8 +369,8 @@ def test_compile_lockstep_raises_for_semantic_errors(
         ],
     )
 
-    with pytest.raises(debug_compiler_module.LockstepCompileError) as exc_info:
-        debug_compiler_module.compile_lockstep("pipeline P { }", verbose=False)
+    with pytest.raises(lockstep_frontend_module.LockstepCompileError) as exc_info:
+        lockstep_frontend_module.compile_lockstep("pipeline P { }", verbose=False)
 
     assert str(exc_info.value) == (
         "Compilation failed with 1 semantic error.\nline 5:2 semantic problem"
@@ -375,7 +378,7 @@ def test_compile_lockstep_raises_for_semantic_errors(
 
 
 def test_compile_lockstep_normalizes_diagnostics_order_and_duplicates(
-    debug_compiler_module, monkeypatch
+    lockstep_frontend_module, monkeypatch
 ):
     class SuccessParser:
         def __init__(self, stream):
@@ -402,14 +405,14 @@ def test_compile_lockstep_normalizes_diagnostics_order_and_duplicates(
             self.uniforms = []
             self.bind_routes = []
             self.diagnostics = [
-                debug_compiler_module.LockstepDiagnostic(
+                lockstep_frontend_module.LockstepDiagnostic(
                     severity="warning",
                     code="LCK099",
                     message="same issue",
                     line=3,
                     column=1,
                 ),
-                debug_compiler_module.LockstepDiagnostic(
+                lockstep_frontend_module.LockstepDiagnostic(
                     severity="error",
                     code="LCK200",
                     message="later diagnostic",
@@ -422,29 +425,29 @@ def test_compile_lockstep_normalizes_diagnostics_order_and_duplicates(
             return tree
 
     monkeypatch.setattr(
-        debug_compiler_module, "CommonTokenStream", lambda lexer: object()
+        lockstep_frontend_module, "CommonTokenStream", lambda lexer: object()
     )
-    monkeypatch.setattr(debug_compiler_module, "LockstepParser", SuccessParser)
-    monkeypatch.setattr(debug_compiler_module, "LockstepDebugVisitor", SpyVisitor)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepParser", SuccessParser)
+    monkeypatch.setattr(lockstep_frontend_module, "LockstepDebugVisitor", SpyVisitor)
     monkeypatch.setattr(
-        debug_compiler_module,
+        lockstep_frontend_module,
         "validate_semantics",
         lambda _parse_tree: [
-            debug_compiler_module.LockstepDiagnostic(
+            lockstep_frontend_module.LockstepDiagnostic(
                 severity="warning",
                 code="LCK099",
                 message="same issue",
                 line=3,
                 column=1,
             ),
-            debug_compiler_module.LockstepDiagnostic(
+            lockstep_frontend_module.LockstepDiagnostic(
                 severity="info",
                 code="LCK050",
                 message="same location info",
                 line=3,
                 column=1,
             ),
-            debug_compiler_module.LockstepDiagnostic(
+            lockstep_frontend_module.LockstepDiagnostic(
                 severity="info",
                 code="LCK010",
                 message="same location info from semantics",
@@ -454,7 +457,7 @@ def test_compile_lockstep_normalizes_diagnostics_order_and_duplicates(
         ],
     )
 
-    result = debug_compiler_module.compile_lockstep("pipeline P { }", verbose=False)
+    result = lockstep_frontend_module.compile_lockstep("pipeline P { }", verbose=False)
 
     assert [
         (d.line, d.column, d.severity, d.code, d.message) for d in result.diagnostics
@@ -474,8 +477,8 @@ def _ctx(start_line=0, start_col=0, **kwargs):
     return types.SimpleNamespace(start=types.SimpleNamespace(line=start_line, column=start_col), **kwargs)
 
 
-def test_visitor_methods_print_expected_output(debug_compiler_module, capsys):
-    visitor = debug_compiler_module.LockstepDebugVisitor()
+def test_visitor_methods_print_expected_output(lockstep_frontend_module, capsys):
+    visitor = lockstep_frontend_module.LockstepDebugVisitor()
 
     class _Param:
         def __init__(self, modifier, p_type, p_name):
@@ -562,8 +565,8 @@ def test_visitor_methods_print_expected_output(debug_compiler_module, capsys):
     assert visitor.diagnostics == []
 
 
-def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_module):
-    visitor = debug_compiler_module.LockstepDebugVisitor(verbose=False)
+def test_visitor_emits_diagnostics_for_non_fatal_observations(lockstep_frontend_module):
+    visitor = lockstep_frontend_module.LockstepDebugVisitor(verbose=False)
 
     visitor.visitStructDecl(_ctx(start_line=2, start_col=1, ID=lambda: _token("Vec3")))
     visitor.visitStructDecl(_ctx(start_line=3, start_col=1, ID=lambda: _token("Vec3")))
@@ -584,7 +587,7 @@ def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_mod
     visitor.visitBindBlock(_ctx(start_line=10, start_col=4, bindStmt=lambda: []))
 
     assert visitor.diagnostics == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="warning",
             code="LCK201",
             message="Struct 'Vec3' is redeclared.",
@@ -592,7 +595,7 @@ def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_mod
             column=1,
             hint="Rename or remove duplicate struct declarations.",
         ),
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="warning",
             code="LCK205",
             message="Pure function 'add' is redeclared.",
@@ -600,7 +603,7 @@ def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_mod
             column=1,
             hint="Rename or remove duplicate pure function declarations.",
         ),
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="warning",
             code="LCK206",
             message="Filter 'f' is redeclared.",
@@ -608,7 +611,7 @@ def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_mod
             column=1,
             hint="Rename or remove duplicate filter declarations.",
         ),
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="warning",
             code="LCK207",
             message="Uniform 'dt' is redeclared.",
@@ -616,7 +619,7 @@ def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_mod
             column=1,
             hint="Each uniform in a pipeline should have a unique name.",
         ),
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="info",
             code="LCK101",
             message="Bind block is empty; pipeline has no executable routes.",
@@ -627,15 +630,15 @@ def test_visitor_emits_diagnostics_for_non_fatal_observations(debug_compiler_mod
     ]
 
 
-def test_visitor_shader_decl_without_param_list(debug_compiler_module, capsys):
-    visitor = debug_compiler_module.LockstepDebugVisitor()
+def test_visitor_shader_decl_without_param_list(lockstep_frontend_module, capsys):
+    visitor = lockstep_frontend_module.LockstepDebugVisitor()
     visitor.visitShaderDecl(_ctx(ID=lambda: _token("Kernel"), paramList=lambda: None))
     assert "[Shader Kernel] Kernel" in capsys.readouterr().out
     assert visitor.shaders == [{"name": "Kernel", "params": []}]
 
 
-def test_visitor_bind_routes_can_be_normalized(debug_compiler_module):
-    visitor = debug_compiler_module.LockstepDebugVisitor(
+def test_visitor_bind_routes_can_be_normalized(lockstep_frontend_module):
+    visitor = lockstep_frontend_module.LockstepDebugVisitor(
         verbose=False,
         normalize_bind_routes=True,
     )
@@ -652,8 +655,8 @@ def test_visitor_bind_routes_can_be_normalized(debug_compiler_module):
     assert visitor.bind_routes == ["a =b(c , d ) ;"]
 
 
-def test_visitor_can_run_without_printing(debug_compiler_module, capsys):
-    visitor = debug_compiler_module.LockstepDebugVisitor(verbose=False)
+def test_visitor_can_run_without_printing(lockstep_frontend_module, capsys):
+    visitor = lockstep_frontend_module.LockstepDebugVisitor(verbose=False)
     visitor.visitStructDecl(_ctx(ID=lambda: _token("Vec3")))
 
     assert capsys.readouterr().out == ""
@@ -662,12 +665,15 @@ def test_visitor_can_run_without_printing(debug_compiler_module, capsys):
 
 def test_module_main_success_path(monkeypatch, capsys):
     _install_fake_generated_modules(monkeypatch)
-    sys.modules.pop("debug_compiler", None)
-    monkeypatch.setattr(sys, "argv", ["debug_compiler.py"])
+    sys.modules.pop("lockstep_frontend", None)
+    sys.modules.pop("lockstep_compiler", None)
+    sys.modules.pop("lockstep_compiler.visitors", None)
+    sys.modules.pop("lockstep_compiler.diagnostics", None)
+    monkeypatch.setattr(sys, "argv", ["lockstep_frontend.py"])
     monkeypatch.setattr(sys, "stdin", io.StringIO("pipeline P { }"))
 
     with pytest.raises(SystemExit) as exc_info:
-        runpy.run_module("debug_compiler", run_name="__main__")
+        runpy.run_module("lockstep_frontend", run_name="__main__")
 
     assert exc_info.value.code == 0
     assert capsys.readouterr().err == ""
@@ -676,12 +682,15 @@ def test_module_main_success_path(monkeypatch, capsys):
 def test_module_main_error_path_exits_with_stderr(monkeypatch, capsys):
     _install_fake_generated_modules(monkeypatch)
     sys.modules["LockstepParser"].LockstepParser.error_to_emit = (7, 9, "bad syntax")
-    sys.modules.pop("debug_compiler", None)
-    monkeypatch.setattr(sys, "argv", ["debug_compiler.py"])
+    sys.modules.pop("lockstep_frontend", None)
+    sys.modules.pop("lockstep_compiler", None)
+    sys.modules.pop("lockstep_compiler.visitors", None)
+    sys.modules.pop("lockstep_compiler.diagnostics", None)
+    monkeypatch.setattr(sys, "argv", ["lockstep_frontend.py"])
     monkeypatch.setattr(sys, "stdin", io.StringIO("pipeline P { }"))
 
     with pytest.raises(SystemExit) as exc_info:
-        runpy.run_module("debug_compiler", run_name="__main__")
+        runpy.run_module("lockstep_frontend", run_name="__main__")
 
     assert exc_info.value.code == 1
     stderr = capsys.readouterr().err
@@ -689,13 +698,13 @@ def test_module_main_error_path_exits_with_stderr(monkeypatch, capsys):
     assert "line 7:9 bad syntax" in stderr
 
 
-def test_run_cli_reads_source_from_stdin_when_path_omitted(debug_compiler_module):
+def test_run_cli_reads_source_from_stdin_when_path_omitted(lockstep_frontend_module):
     captured = {}
 
     def fake_compiler(source):
         captured["source"] = source
 
-    exit_code = debug_compiler_module.run_cli(
+    exit_code = lockstep_frontend_module.run_cli(
         [],
         stdin=io.StringIO("pipeline FromStdin { }"),
         compiler=fake_compiler,
@@ -705,7 +714,7 @@ def test_run_cli_reads_source_from_stdin_when_path_omitted(debug_compiler_module
     assert captured["source"] == "pipeline FromStdin { }"
 
 
-def test_run_cli_reads_source_from_path(debug_compiler_module, tmp_path):
+def test_run_cli_reads_source_from_path(lockstep_frontend_module, tmp_path):
     source_file = tmp_path / "sample.lock"
     source_file.write_text("pipeline FromFile { }", encoding="utf-8")
     captured = {}
@@ -713,7 +722,7 @@ def test_run_cli_reads_source_from_path(debug_compiler_module, tmp_path):
     def fake_compiler(source):
         captured["source"] = source
 
-    exit_code = debug_compiler_module.run_cli(
+    exit_code = lockstep_frontend_module.run_cli(
         [str(source_file)], compiler=fake_compiler
     )
 
@@ -721,11 +730,11 @@ def test_run_cli_reads_source_from_path(debug_compiler_module, tmp_path):
     assert captured["source"] == "pipeline FromFile { }"
 
 
-def test_run_cli_returns_non_zero_and_writes_errors(debug_compiler_module):
+def test_run_cli_returns_non_zero_and_writes_errors(lockstep_frontend_module):
     def failing_compiler(_source):
-        raise debug_compiler_module.LockstepCompileError(
+        raise lockstep_frontend_module.LockstepCompileError(
             [
-                debug_compiler_module.LockstepDiagnostic(
+                lockstep_frontend_module.LockstepDiagnostic(
                     severity="error",
                     code="LCK001",
                     message="unexpected",
@@ -737,7 +746,7 @@ def test_run_cli_returns_non_zero_and_writes_errors(debug_compiler_module):
         )
 
     stderr = io.StringIO()
-    exit_code = debug_compiler_module.run_cli(
+    exit_code = lockstep_frontend_module.run_cli(
         [],
         stdin=io.StringIO("pipeline Broken {"),
         stderr=stderr,
@@ -751,7 +760,7 @@ def test_run_cli_returns_non_zero_and_writes_errors(debug_compiler_module):
     ]
 
 
-def test_run_cli_returns_non_zero_for_missing_path(debug_compiler_module, tmp_path):
+def test_run_cli_returns_non_zero_for_missing_path(lockstep_frontend_module, tmp_path):
     missing = tmp_path / "missing.lock"
     stderr = io.StringIO()
     called = {"compiler": False}
@@ -759,7 +768,7 @@ def test_run_cli_returns_non_zero_for_missing_path(debug_compiler_module, tmp_pa
     def fake_compiler(_source):
         called["compiler"] = True
 
-    exit_code = debug_compiler_module.run_cli(
+    exit_code = lockstep_frontend_module.run_cli(
         [str(missing)],
         stderr=stderr,
         compiler=fake_compiler,
@@ -770,7 +779,7 @@ def test_run_cli_returns_non_zero_for_missing_path(debug_compiler_module, tmp_pa
     assert f"Unable to read '{missing}': file not found." in stderr.getvalue()
 
 
-def test_run_cli_returns_non_zero_for_unreadable_path(debug_compiler_module, monkeypatch):
+def test_run_cli_returns_non_zero_for_unreadable_path(lockstep_frontend_module, monkeypatch):
     stderr = io.StringIO()
     called = {"compiler": False}
 
@@ -780,9 +789,9 @@ def test_run_cli_returns_non_zero_for_unreadable_path(debug_compiler_module, mon
     def raise_permission_error(_self, encoding):
         raise PermissionError("permission denied")
 
-    monkeypatch.setattr(debug_compiler_module.Path, "read_text", raise_permission_error)
+    monkeypatch.setattr(lockstep_frontend_module.Path, "read_text", raise_permission_error)
 
-    exit_code = debug_compiler_module.run_cli(
+    exit_code = lockstep_frontend_module.run_cli(
         ["locked.lock"],
         stderr=stderr,
         compiler=fake_compiler,
@@ -793,7 +802,7 @@ def test_run_cli_returns_non_zero_for_unreadable_path(debug_compiler_module, mon
     assert "Unable to read 'locked.lock': permission denied." in stderr.getvalue()
 
 
-def test_run_cli_returns_non_zero_for_invalid_utf8(debug_compiler_module, tmp_path):
+def test_run_cli_returns_non_zero_for_invalid_utf8(lockstep_frontend_module, tmp_path):
     bad_source = tmp_path / "invalid.lock"
     bad_source.write_bytes(b"\xff\xfe\xfa")
     stderr = io.StringIO()
@@ -802,7 +811,7 @@ def test_run_cli_returns_non_zero_for_invalid_utf8(debug_compiler_module, tmp_pa
     def fake_compiler(_source):
         called["compiler"] = True
 
-    exit_code = debug_compiler_module.run_cli(
+    exit_code = lockstep_frontend_module.run_cli(
         [str(bad_source)],
         stderr=stderr,
         compiler=fake_compiler,
@@ -813,8 +822,8 @@ def test_run_cli_returns_non_zero_for_invalid_utf8(debug_compiler_module, tmp_pa
     assert f"Unable to read '{bad_source}': invalid UTF-8" in stderr.getvalue()
 
 
-def test_semantic_validator_reports_undefined_identifier_in_bind(debug_compiler_module):
-    validator = debug_compiler_module.LockstepSemanticValidator()
+def test_semantic_validator_reports_undefined_identifier_in_bind(lockstep_frontend_module):
+    validator = lockstep_frontend_module.LockstepSemanticValidator()
     validator.shaders = {
         "Apply": [
             {"name": "inp", "type": "Vec3", "modifier": "in"},
@@ -835,7 +844,7 @@ def test_semantic_validator_reports_undefined_identifier_in_bind(debug_compiler_
     validator.visitBindStmt(bind_ctx)
 
     assert validator.diagnostics == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="error",
             code="LCK301",
             message="Undefined identifier 'missing_stream'.",
@@ -846,8 +855,8 @@ def test_semantic_validator_reports_undefined_identifier_in_bind(debug_compiler_
     ]
 
 
-def test_semantic_validator_reports_bind_arity_and_type_errors(debug_compiler_module):
-    validator = debug_compiler_module.LockstepSemanticValidator()
+def test_semantic_validator_reports_bind_arity_and_type_errors(lockstep_frontend_module):
+    validator = lockstep_frontend_module.LockstepSemanticValidator()
     validator.shaders = {
         "Apply": [
             {"name": "inp", "type": "Vec3", "modifier": "in"},
@@ -882,8 +891,8 @@ def test_semantic_validator_reports_bind_arity_and_type_errors(debug_compiler_mo
     assert "expected float, got int" in validator.diagnostics[1].message
 
 
-def test_semantic_validator_reports_duplicate_pipeline_symbols(debug_compiler_module):
-    validator = debug_compiler_module.LockstepSemanticValidator()
+def test_semantic_validator_reports_duplicate_pipeline_symbols(lockstep_frontend_module):
+    validator = lockstep_frontend_module.LockstepSemanticValidator()
     validator._push_scope()
 
     duplicate_ctx = _ctx(start_line=7, start_col=1, ID=lambda: _token("energy"), typeName=lambda: _token("float"))
@@ -891,7 +900,7 @@ def test_semantic_validator_reports_duplicate_pipeline_symbols(debug_compiler_mo
     validator.visitUniformDecl(duplicate_ctx)
 
     assert validator.diagnostics == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="error",
             code="LCK306",
             message="Duplicate declaration for 'energy' in the same scope.",
@@ -902,8 +911,8 @@ def test_semantic_validator_reports_duplicate_pipeline_symbols(debug_compiler_mo
     ]
 
 
-def test_semantic_validator_reports_fold_reference_errors(debug_compiler_module):
-    validator = debug_compiler_module.LockstepSemanticValidator()
+def test_semantic_validator_reports_fold_reference_errors(lockstep_frontend_module):
+    validator = lockstep_frontend_module.LockstepSemanticValidator()
     validator._push_scope()
     validator._declare("not_acc", "float", _ctx(), duplicate_code="LCK306", kind="uniform")
     validator._declare("acc_energy", "float", _ctx(), duplicate_code="LCK306", kind="accumulator")
@@ -933,9 +942,9 @@ def test_semantic_validator_reports_fold_reference_errors(debug_compiler_module)
 
 
 def test_semantic_validator_nested_lvalue_reports_single_undefined_identifier(
-    debug_compiler_module,
+    lockstep_frontend_module,
 ):
-    validator = debug_compiler_module.LockstepSemanticValidator()
+    validator = lockstep_frontend_module.LockstepSemanticValidator()
     validator._push_scope()
 
     lvalue_ctx = _ctx(
@@ -962,7 +971,7 @@ def test_semantic_validator_nested_lvalue_reports_single_undefined_identifier(
     validator.visitPrimaryExpr(primary_ctx)
 
     assert validator.diagnostics == [
-        debug_compiler_module.LockstepDiagnostic(
+        lockstep_frontend_module.LockstepDiagnostic(
             severity="error",
             code="LCK301",
             message="Undefined identifier 'missing'.",
